requirements.txt
requests
codecov
gitpython
pre-commit
pytest
pytest-cov
pyyaml
quack
wget
argparse
colorama
humanfriendly
phonenumbers
scapy
python-dotenv
discord.py==1.6.0
